requirements.txt
flask
flask-socketio
flask-sqlalchemy
flask-migrate
waitress
requests
mcstatus
pyjwt
flask-bcrypt
flask-limiter